Class FLXSA

A class for the results of an XSA analysis.

Objects from the Class

Objects can be created by calls of the form new("FLXSA", ...) and are output by calls to FLXSA.

Slots

survivors:

Object of class "FLQuant". Numbers-at-age in the final age of a cohort

se.int:

Object of class "FLQuant". Internal standard errors of numbers-at-age in the final year

se.ext:

Object of class "FLQuant". External standard errors of numbers-at-age in the final year

n.fshk:

Object of class "FLQuant". Numbers-at-age in the final age of a cohort derived from shrinkage to the mean F

n.nshk:

Object of class "FLQuant". Numbers-at-age in the final age of a cohort derived from shrinkage to the mean N

var.fshk:

Object of class "FLQuant". Numbers-at-age in the final age of a cohort derived from shrinkage to the mean F

var.nshk:

Object of class "FLQuant". Variance assumed for numbers-at-age in the final age of a cohort derived from shrinkage to the mean N

q.hat:

Object of class "list". Estimates of q-at-age by cpue index

q2.hat:

Object of class "list". Estimates of exponent in cpue power model

control:

Object of class "FLXSA.control". FLXSA.control object used.

catch.n:

Object of class "FLQuant". Estimated catch numbers-at-age, not estimated

stock.n:

Object of class "FLQuant". Estimated stock numbers-at-age

harvest:

Object of class "FLQuant". Estimated catch F-at-age

index.name:

Object of class "character". Names of cpue indices

index:

Object of class "list". cpue indices, adjusted to start of year

index.res:

Object of class "list". cpue index residuals

index.hat:

Object of class "list". predicted cpue indices

index.var:

Object of class "list". Estimated variance of cpue index residuals

call:

A "character" representation of the call that created the object

desc:

A description on the object, of class "character"

diagnostics:

A data frame with various diagnostics from a run of XSA.

Extends

Class "FLAssess", directly.

Methods

All methods are inherited from FLAssess.

References

Darby, C. D., and Flatman, S. 1994. Virtual Population Analysis: version 3.1 (Windows/Dos) user guide. Info. Tech. Ser., MAFF Direct. Fish. Res., Lowestoft, (1): 85pp.

Shepherd, J.G. 1992. Extended survivors analysis: an improved method for the analysis of catch-at-age data and catch-per-unit-effort data. Working paper No. 11 ICES Multi-species Assessment Working Group, June 1992, Copenhagen, Denmark. 22pp. (mimeo).

Shepherd, J.G. 1994. Prediction of yearclass strength by calibration regression analysis of multiple recruit index series. ICES J. Mar. Sci. In Prep.

See also

FLXSA

Examples