Class "FLXSA.control"

control object for the XSA (eXtended Survivors Analysis) model.

Objects from the Class

Objects can be created by calls of the form new("FLXSA.control", ...) or through the creator function FLXSA.control.

Slots

tol:

The covergence tolerance, i.e. difference between two successive iterations must be lower to declare convergence of the model. Class "numeric"

maxit:

The maximum number of iterations allowed. Class "integer"

min.nse:

The minimum value of SE permitted in estimate of N hat. Class "numeric"

fse:

User set SE of F when shrinking to mean F. Class "numeric"

rage:

The oldest age for which the two parameter model is used for catchability at age. Class "integer"

qage:

The age after which catchability is no longer estimated. Class "integer"

shk.n:

If TRUE, shrinkage to mean N. Class "logical"

shk.f:

If TRUE, shrinkage to mean F. Class "logical"

shk.yrs:

The number of years to be used for shrinkage to F for terminal year. Class "integer"

shk.ages:

The number of ages to be used for shrinkage to F for terminal age. Class "integer"

window:

The time window to consider in the model fitting. Class "integer"

tsrange:

The number of years to be used in the time series weighting. Class "integer" ~~

tspower:

The power to be used in the time series taper weighting. Class "integer"

vpa:

If FALSE use cohort analysis, otherwise, use VPA. Class "logical"

Methods

assess

signature(object = "FLXSA.control"): Runs FLXSA according to the options defined in the object.

See also

FLXSA, FLXSA-class